optimize datasource usage of driver (#207)
optimize: optimize datasource usage of driver
